server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
structlog==23.1.0
Flask-CORS==4.0.1